Fix menu search keyboard (#637)
Fixes https://github.com/datagouv/data.gouv.fr/issues/1640

Keyboard navigation was broken on the menu search and probably other
components too.

We were getting a list of keycodes (number representing the `ArrowDown`
key for example) from the DSFR because they were already defined but it
seems that the existing object changed to a nested one... So It's now a
simple type instead.